refactor(canvas): move trackAndCreateCanvas out of the component file
Fixes the React Doctor only-export-components warning on NewCanvasMenu.tsx by moving the CreateSurface type and trackAndCreateCanvas helper to a plain module.

Generated-By: PostHog Code
Task-Id: e556b691-318b-4472-8771-d397fc3d221d
